+++ src/common/chapters/chapters.h
@@ -84,7 +84,7 @@ libmatroska::KaxEditionEntry *find_edition_with_uid(libmatroska::KaxChapters &ch
 libmatroska::KaxChapterAtom *find_chapter_with_uid(libmatroska::KaxChapters &chapters, uint64_t uid);
 
 void move_by_edition(libmatroska::KaxChapters &dst, libmatroska::KaxChapters &src);
-void adjust_timestamps(libebml::EbmlMaster &master, int64_t offset);
+void adjust_timestamps(libebml::EbmlMaster &master, int64_t offset, int64_t numerator = 1, int64_t denominator = 1);
 void merge_entries(libebml::EbmlMaster &master);
 int count_atoms(libebml::EbmlMaster &master);
 void regenerate_uids(libebml::EbmlMaster &master);